rename tool utils
