+++ auto-agents-framework/src/agents/tools/utils/agentWallet.ts
