+++ auto-agents-framework/src/agents/workflows/kol/nodes/collectDataNode.ts
@@ -1,7 +1,7 @@
 import { WorkflowConfig } from '../types.js';
 import { createLogger } from '../../../../utils/logger.js';
 import { State } from '../workflow.js';
-import { convertMessageContentToTweets } from '../../../tools/utils/twitterToolUtils.js';
+import { convertMessageContentToTweets } from '../../../tools/utils/twitter.js';
 import { invokeFetchTimelineTool } from '../../../tools/fetchTimelineTool.js';
 import { invokeFetchMentionsTool } from '../../../tools/fetchMentionsTool.js';
